Add a generic table function, add support for LenT
